Remove legacy pre-notifications-stream announcement support.
Before we added the notifications stream feature, Zulip used to send
PMs to all users in the realm to announce new streams.  We left in the
code for the old PM feature when we initially added the notifications
stream feature intended to be for a transitional period, and at this
point it seems reasonable to remove it.