Add Attachment model to keep track of uploads.
This commit adds the capability to keep track and remove uploaded files. Unclaimed
attachments are files that have been uploaded to the server but are not referred
in any messages. A management command to remove old unclaimed files is also included.

Tests for getting the file referred in messages are also included.