Add teamcity webhook integration.
This integration relies on the Teamcity "tcWebHooks" plugin which is
available at
https://netwolfuk.wordpress.com/category/teamcity/tcplugins/tcwebhooks/

It posts build fail and success notifications to a stream specified in
the webhook.
It uses the name of the build configuration as the topic.

For personal builds, it tries to map the Teamcity username to a Zulip username, and private messages that person.