Fix issue #853 add warning for @all / @everyone
The warning contains a count of the number of people in the stream.
An error appears if the warning is ignored and the user tries to send the message anyway.
The message cannot be sent until the warning is acknowledged or @all / @everyone is removed.
This only applies to stream messages and not private messages.