Make the "M"/"U" keys mute/unmute a topic
Being in streams that generate a lot of topics that have a medium chance
of being uninteresting can make it hard to keep up - muting a topic can
help, but without a key binding, it's pretty tedious.

This change adds the "M" key binding that allows muting a topic very
quickly, and an "U" key binding that allows unmuting it from the topic's
stream.