uploads: Add authorization check before serving files.
This is a remerge of e985b57259f86dcdbddb115fc5009e2db04496ff (after
resolving merge conflicts, updating the tests, adding mypy annotations
etc.), which should now be correct, because we've done the necessary
database migration.

The rebase/remerge work was done by Tim Abbott and Aditya Bansal.

This is an important part of #320.