Add subject_topic_awareness() test helper.
This is a purely tactical commit that adds a context
manager called subject_topic_awareness().  It is intended
to reduce merge conflicts when we start moving over to a
world where topics are stored in a Topic table.

All tests that use this context manager do meaningful
writes/edits to the zerver_message table that touch
Message.subject.