Set up v1 of analytics framework.
This is a first pass at building a framework for collecting various
stats about realms, users, streams etc. We allow a user to add a
CountStat object that will pull various counts and their aggregations
via a command that are queried from the prod database and aggregated.
New analytics_* tables are added to the model
management command added to process counts for a CountStat object
zerver pull function added to query the prod db
interval functions added
aggregate functions added to aggregate counts based on interval
data_collectors added to exclude list so that linter does not throw % formatting errors.