Split out docs/zjsunit.md from docs/testing.md
This commit also adds a small intro to 'docs/testing.md'.

It also removes a statement that preceded this change that
had a broken link, and which was slightly out of context anyway.
(I removed it in this commit to avoid confusion for somebody who
may have thought this change broke the link.)