ModelReprMixin: Return string error message when __unicode__ is not implemented.
Old behavior was to raise an exception, but Django was catching the
exception and doing unexpected things. For instance, in the manage.py shell,
printing out a ModelReprMixin object (with __unicode__ not implemented)
would result in nothing being printed, rather than it raising a error or
otherwise alerting the programmer as to what was going on.