Adding required styling for displaying user information in the navbar.
This is the requisite styling for implementing the full name and
organization name in the right and left columns of the navbar
respectively.

These values truncate to ellipses if too big to fit their containers
and they disappear at different widths.

The full name on the left disappears at 975px wide and below, and the
organization name disappears at 775px.